requirements.txt
# Core Framework
python-telegram-bot==20.7
pandas>=1.5.0
numpy>=1.21.0
pydantic>=1.10.0
pyyaml>=6.0

# MT5 Integration
MetaTrader5>=5.0.35

# Technical Analysis & Charting
mplfinance>=0.12.0
ta-lib>=0.4.24

# Web & Data
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Utilities
python-json-logger>=2.0.0
pytz>=2022.0
python-dateutil>=2.8.0

# Advanced Analytics
scipy>=1.7.0
statsmodels>=0.13.0